+++ sofa-tracer-plugins/sofa-tracer-datasource-plugin/src/test/java/com/sofa/tracer/plugins/datasource/DataSourceUtilTest.java
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.sofa.tracer.plugins.datasource;
+
+import com.alipay.common.tracer.core.utils.ReflectionUtils;
+import com.sofa.tracer.plugins.datasource.bean.ConcreteClassService;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.lang.reflect.Method;
+
+/**
+ * @author qilong.zql
+ * @since 2.3.2
+ */
+public class DataSourceUtilTest {
+    @Test
+    public void testReflectionUtils() throws Throwable {
+        ConcreteClassService concreteClassService = new ConcreteClassService();
+        Method method = ReflectionUtils.findMethod(concreteClassService.getClass(), "service");
+        Assert.assertEquals("concreteClassService", method.invoke(concreteClassService));
+        method = ReflectionUtils.findMethod(concreteClassService.getClass(), "serviceA");
+        Assert.assertEquals("serviceA", method.invoke(concreteClassService));
+        method = ReflectionUtils.findMethod(concreteClassService.getClass(), "serviceB");
+        Assert.assertEquals("serviceB", method.invoke(concreteClassService));
+    }
+}
