+++ sofa-tracer-plugins/sofa-tracer-datasource-plugin/src/test/java/com/sofa/tracer/plugins/datasource/bean/InterfaceB.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.tracer.boot.datasource.bean;
+package com.sofa.tracer.plugins.datasource.bean;
 
 /**
  * @author qilong.zql
